MNT: move ExpandableFrame to top-level atef.widgets.utils
